Move functions around to different files
