+++ pkg/agents/team5/agent.go
@@ -88,97 +88,6 @@ func (a *CustomAgent5) updateSatisfaction() {
 	}
 }
 
-func (a *CustomAgent5) checkForLeader() {
-	// Random number between 3 and 12 generated, then the agent floor and selfishness are deducted from this
-	diceRoll := rand.Intn(10) + 3 - a.selfishness - a.Floor()
-	if diceRoll >= a.leadership {
-		a.Log("An agent has become a leader", infra.Fields{"dice roll": diceRoll, "leadership": a.leadership, "selfishness": a.selfishness, "floor": a.Floor()})
-		//TODO: Send treaties here about eating less food
-		hpLevel := a.currentAimHP - ((a.currentAimHP-a.HealthInfo().WeakLevel)/10)*(diceRoll-a.leadership)
-		a.currentProposal = messages.NewTreaty(messages.HP, hpLevel, messages.LeavePercentFood, 100, messages.GE, messages.EQ, 5, a.ID())
-		a.treatySendCounter = 1
-		a.Log("Agent is sending a treaty proposal", infra.Fields{"Proposed Max Hp": hpLevel})
-		a.currentProposal.SignTreaty()
-		a.AddTreaty(*a.currentProposal)
-	}
-}
-
-func (a *CustomAgent5) treatiesCanCoexist(t1 messages.Treaty, t2 messages.Treaty) bool {
-	conditionsIntersect := statementsIntersect(t1.ConditionOp(), t1.ConditionValue(), t2.ConditionOp(), t2.ConditionValue())
-
-	// If the treaties are based on the same type of condition, they can coexist
-	// when the conditions are mutually exclusive (can never occur simultaneously)
-	if t1.Condition() == t2.Condition() && !conditionsIntersect {
-		return true
-	}
-
-	// Otherwise, if the treaties are based on different conditions, or the conditions
-	// can occurr simultaneously, treaties of different types should be rejected
-	// because it is impossible to deduce if their requests can be fulfilled simultaneously
-	if t1.Request() != t2.Request() {
-		return false
-	}
-
-	// Otherwise, treaties can coexist only if the requests can be fulfilled simultaneously
-	return statementsIntersect(t1.RequestOp(), t1.RequestValue(), t2.RequestOp(), t2.RequestValue())
-}
-
-func (a *CustomAgent5) treatyConflicts(treaty messages.Treaty) bool {
-	for _, activeTreaty := range a.ActiveTreaties() {
-		if !a.treatiesCanCoexist(treaty, activeTreaty) {
-			return true
-		}
-	}
-	return false
-}
-
-func (a *CustomAgent5) overrideCalculation(treaty messages.Treaty) {
-	if treaty.Request() == messages.LeaveAmountFood {
-		if treaty.RequestOp() == messages.GT && a.CurrPlatFood()-a.attemptFood < food.FoodType(treaty.RequestValue()) {
-			a.attemptFood = food.FoodType(math.Max(0, float64(a.CurrPlatFood()-food.FoodType(treaty.RequestValue())-1)))
-		}
-		if treaty.RequestOp() == messages.GE && a.CurrPlatFood()-a.attemptFood <= food.FoodType(treaty.RequestValue()) {
-			a.attemptFood = food.FoodType(math.Max(0, float64(a.CurrPlatFood()-food.FoodType(treaty.RequestValue()))))
-		}
-	}
-	if treaty.Request() == messages.LeavePercentFood {
-		if treaty.RequestOp() == messages.EQ {
-			a.attemptFood = a.CurrPlatFood() - food.FoodType(float64(a.CurrPlatFood())*float64(treaty.RequestValue())/100)
-		}
-		if treaty.RequestOp() == messages.GT && a.CurrPlatFood()-a.attemptFood < food.FoodType(float64(a.CurrPlatFood())*float64(treaty.RequestValue()/100)) {
-			a.attemptFood = a.CurrPlatFood() - food.FoodType(float64(a.CurrPlatFood())*float64(treaty.RequestValue())/100) - 1
-		}
-		if treaty.RequestOp() == messages.GE && a.CurrPlatFood()-a.attemptFood <= food.FoodType(float64(a.CurrPlatFood())*float64(treaty.RequestValue()/100)) {
-			a.attemptFood = a.CurrPlatFood() - food.FoodType(float64(a.CurrPlatFood())*float64(treaty.RequestValue())/100)
-		}
-	}
-}
-
-func (a *CustomAgent5) treatyOverride() {
-	for _, treaty := range a.ActiveTreaties() {
-		switch {
-		case treaty.Condition() == messages.HP && treaty.ConditionOp() == messages.EQ && a.HP() == treaty.ConditionValue():
-			fallthrough
-		case treaty.Condition() == messages.HP && treaty.ConditionOp() == messages.GT && a.HP() > treaty.ConditionValue():
-			fallthrough
-		case treaty.Condition() == messages.HP && treaty.ConditionOp() == messages.GE && a.HP() >= treaty.ConditionValue():
-			fallthrough
-		case treaty.Condition() == messages.Floor && treaty.ConditionOp() == messages.EQ && a.Floor() == treaty.ConditionValue():
-			fallthrough
-		case treaty.Condition() == messages.Floor && treaty.ConditionOp() == messages.LT && a.Floor() < treaty.ConditionValue():
-			fallthrough
-		case treaty.Condition() == messages.Floor && treaty.ConditionOp() == messages.LE && a.Floor() <= treaty.ConditionValue():
-			fallthrough
-		case treaty.Condition() == messages.AvailableFood && treaty.ConditionOp() == messages.EQ && a.CurrPlatFood() == food.FoodType(treaty.ConditionValue()):
-			fallthrough
-		case treaty.Condition() == messages.AvailableFood && treaty.ConditionOp() == messages.GT && a.CurrPlatFood() > food.FoodType(treaty.ConditionValue()):
-			fallthrough
-		case treaty.Condition() == messages.AvailableFood && treaty.ConditionOp() == messages.GE && a.CurrPlatFood() >= food.FoodType(treaty.ConditionValue()):
-			a.overrideCalculation(treaty)
-		}
-	}
-}
-
 func (a *CustomAgent5) dayPassed() {
 	a.updateFavour()
 	a.updateSelfishness()
